--- src/main/java/com/team/independence/asset/service/AssetAccountListService.java
@@ -1,7 +1,0 @@
-package com.team.independence.asset.service;
-
-import com.team.independence.asset.dto.AssetAccountListResponse;
-
-public interface AssetAccountListService {
-    AssetAccountListResponse getAccountList(Long memberId);
-}
