+++ src/main/java/com/team/independence/asset/service/AssetConnectionServiceImpl.java
@@ -1,16 +1,13 @@
 package com.team.independence.asset.service;
 
-import com.team.independence.asset.domain.AssetSummary;
-import com.team.independence.asset.domain.ConnectedAccount;
-import com.team.independence.asset.domain.ConnectedInstitution;
-import com.team.independence.asset.dto.AssetAccountQueryItem;
-import com.team.independence.asset.dto.AssetLinkRequest;
-import com.team.independence.asset.dto.AssetLinkResponse;
-import com.team.independence.asset.dto.AssetNetWorthBreakdown;
-import com.team.independence.asset.dto.CardAccountResponse;
-import com.team.independence.asset.dto.LinkedOrganizationResponse;
-import com.team.independence.asset.dto.UnlinkOrganizationResponse;
-import com.team.independence.asset.domain.Institution;
+import com.team.independence.asset.domain.codef.ConnectedAccount;
+import com.team.independence.asset.domain.codef.ConnectedInstitution;
+import com.team.independence.asset.domain.codef.Institution;
+import com.team.independence.asset.domain.summary.AssetSummary;
+import com.team.independence.asset.dto.connection.AssetLinkRequest;
+import com.team.independence.asset.dto.connection.AssetLinkResponse;
+import com.team.independence.asset.dto.connection.LinkedOrganizationResponse;
+import com.team.independence.asset.dto.connection.UnlinkOrganizationResponse;
 import com.team.independence.asset.mapper.AssetAccountMapper;
 import com.team.independence.asset.mapper.AssetSummaryMapper;
 import com.team.independence.asset.mapper.CardAccountMapper;
@@ -18,7 +15,6 @@ import com.team.independence.asset.mapper.ConnectedAccountMapper;
 import com.team.independence.asset.mapper.ConnectedInstitutionMapper;
 import com.team.independence.asset.mapper.InstitutionMapper;
 import com.team.independence.asset.mapper.LoanAccountMapper;
-import com.team.independence.asset.mapper.ManualAssetMapper;
 import com.team.independence.common.exception.BusinessException;
 import com.team.independence.common.exception.ErrorCode;
 import com.team.independence.common.security.AesEncryptor;
@@ -41,13 +37,10 @@ import java.util.concurrent.TimeUnit;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class AssetServiceImpl implements AssetService {
+public class AssetConnectionServiceImpl implements AssetConnectionService {
 
     private static final String LOCK_KEY_PREFIX = "asset:link:lock:";
     private static final long LOCK_TTL_SECONDS = 30L;
-
-    /** STOCK/FUND 계좌 평가금액 인정률. 자산 요약 API(AssetSummaryServiceImpl)와 동일 기준. */
-    private static final double INVESTMENT_RECOGNITION_RATE = 0.7;
 
     private final ConnectedAccountMapper connectedAccountMapper;
     private final ConnectedInstitutionMapper connectedInstitutionMapper;
@@ -56,7 +49,6 @@ public class AssetServiceImpl implements AssetService {
     private final AssetSummaryMapper assetSummaryMapper;
     private final CardAccountMapper cardAccountMapper;
     private final LoanAccountMapper loanAccountMapper;
-    private final ManualAssetMapper manualAssetMapper;
     private final CodefClient codefClient;
     private final CodefTokenManager codefTokenManager;
     private final CodefProperties codefProperties;
@@ -192,7 +184,6 @@ public class AssetServiceImpl implements AssetService {
                 .loginType(institutionInfo.getLoginType())
                 .build();
 
-        // CODEF 삭제 성공 후에만 DB 삭제
         codefClient.deleteAccount(accessToken, connectedId, item);
 
         assetAccountMapper.deleteByConnectedInstitutionId(institution.getId());
@@ -221,58 +212,9 @@ public class AssetServiceImpl implements AssetService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<CardAccountResponse> getCardList(Long memberId) {
-        return cardAccountMapper.findByMemberId(memberId).stream()
-                .map(CardAccountResponse::from)
-                .toList();
-    }
-
-    @Override
-    @Transactional(readOnly = true)
     public void validateConnectedAccountExists(Long memberId) {
         if (connectedAccountMapper.findByMemberId(memberId) == null) {
             throw new BusinessException(ErrorCode.ASSET_NOT_LINKED);
         }
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public AssetNetWorthBreakdown getNetWorthBreakdown(Long memberId) {
-        long interestBearingAssets = assetAccountMapper.sumCurrentValueByMemberIdAndCategories(
-                memberId, List.of("DEPOSIT_SAVINGS"));
-        long investmentRecognized = investmentRecognizedAmount(memberId);
-        long cashAndEtcAssets = assetAccountMapper.sumCurrentValueByMemberIdAndCategories(
-                memberId, List.of("CASH", "ETC"));
-        long manualAssets = manualAssetMapper.sumAmountByMemberId(memberId);
-        long loanBalance = loanAccountMapper.sumLoanBalanceByMemberId(memberId);
-
-        long flatRecognizedAssets = investmentRecognized + cashAndEtcAssets + manualAssets - loanBalance;
-
-        return AssetNetWorthBreakdown.builder()
-                .interestBearingAssets(interestBearingAssets)
-                .flatRecognizedAssets(flatRecognizedAssets)
-                .build();
-    }
-
-    /**
-     * INVESTMENT 계좌 인정액 합계.
-     * STOCK/FUND(평가금액 있는 계좌)는 평가금액×70% + 입금대기금, 그 외(CMA 등)는 current_value 그대로.
-     * 자산 요약 API(AssetSummaryServiceImpl.computeBalance)와 동일한 계산 기준을 따른다.
-     */
-    private long investmentRecognizedAmount(Long memberId) {
-        long total = 0L;
-        for (AssetAccountQueryItem account : assetAccountMapper.findWithInstitutionByMemberId(memberId)) {
-            if (!"INVESTMENT".equals(account.getAssetCategory())) {
-                continue;
-            }
-            String type = account.getAccountType();
-            if (("STOCK".equals(type) || "FUND".equals(type)) && account.getValuationAmount() != null) {
-                long deposit = account.getDepositReceived() != null ? account.getDepositReceived() : 0L;
-                total += Math.round(account.getValuationAmount() * INVESTMENT_RECOGNITION_RATE) + deposit;
-            } else {
-                total += account.getCurrentValue() != null ? account.getCurrentValue() : 0L;
-            }
-        }
-        return total;
     }
 }
