+++ src/main/java/com/team/independence/asset/service/AssetSyncService.java
@@ -1,7 +1,7 @@
 package com.team.independence.asset.service;
 
-import com.team.independence.asset.dto.AssetSyncResponse;
-import com.team.independence.asset.dto.SyncJobStatusResponse;
+import com.team.independence.asset.dto.sync.AssetSyncResponse;
+import com.team.independence.asset.dto.sync.SyncJobStatusResponse;
 
 public interface AssetSyncService {
     AssetSyncResponse syncAccounts(Long memberId);
