+++ src/main/java/com/team/independence/asset/service/AssetSyncServiceImpl.java
@@ -2,22 +2,22 @@ package com.team.independence.asset.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.team.independence.asset.domain.AssetAccount;
-import com.team.independence.asset.domain.CardAccount;
-import com.team.independence.asset.domain.ConnectedAccount;
-import com.team.independence.asset.domain.ConnectedInstitution;
-import com.team.independence.asset.domain.LoanAccount;
-import com.team.independence.asset.dto.AssetSyncResponse;
-import com.team.independence.asset.dto.AssetSyncResponse.InstitutionSyncResult;
-import com.team.independence.asset.dto.SyncJobStatusResponse;
-import com.team.independence.asset.domain.AssetSummary;
+import com.team.independence.asset.domain.account.AssetAccount;
+import com.team.independence.asset.domain.account.CardAccount;
+import com.team.independence.asset.domain.codef.ConnectedAccount;
+import com.team.independence.asset.domain.codef.ConnectedInstitution;
+import com.team.independence.asset.domain.account.LoanAccount;
+import com.team.independence.asset.dto.sync.AssetSyncResponse;
+import com.team.independence.asset.dto.sync.AssetSyncResponse.InstitutionSyncResult;
+import com.team.independence.asset.dto.sync.SyncJobStatusResponse;
+import com.team.independence.asset.domain.summary.AssetSummary;
 import com.team.independence.asset.mapper.AssetAccountMapper;
 import com.team.independence.asset.mapper.AssetSummaryMapper;
 import com.team.independence.asset.mapper.CardAccountMapper;
 import com.team.independence.asset.mapper.ConnectedAccountMapper;
 import com.team.independence.asset.mapper.ConnectedInstitutionMapper;
 import com.team.independence.asset.mapper.InstitutionMapper;
-import com.team.independence.asset.domain.Institution;
+import com.team.independence.asset.domain.codef.Institution;
 import com.team.independence.asset.mapper.LoanAccountMapper;
 import com.team.independence.common.exception.BusinessException;
 import com.team.independence.common.exception.ErrorCode;
