+++ src/main/java/com/team/independence/asset/service/BankAccountService.java
@@ -1,0 +1,9 @@
+package com.team.independence.asset.service;
+
+import com.team.independence.asset.dto.account.AssetAccountDetailItem;
+
+import java.util.List;
+
+public interface BankAccountService {
+    List<AssetAccountDetailItem> getBankAccounts(Long memberId);
+}
