+++ src/main/java/com/team/independence/asset/service/BankAccountServiceImpl.java
@@ -1,0 +1,22 @@
+package com.team.independence.asset.service;
+
+import com.team.independence.asset.dto.account.AssetAccountDetailItem;
+import com.team.independence.asset.mapper.AssetAccountMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class BankAccountServiceImpl implements BankAccountService {
+
+    private final AssetAccountMapper assetAccountMapper;
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<AssetAccountDetailItem> getBankAccounts(Long memberId) {
+        return assetAccountMapper.findBankAccountDetailsByMemberId(memberId);
+    }
+}
