+++ src/main/java/com/team/independence/asset/service/CardAccountService.java
@@ -1,0 +1,9 @@
+package com.team.independence.asset.service;
+
+import com.team.independence.asset.dto.account.CardAccountResponse;
+
+import java.util.List;
+
+public interface CardAccountService {
+    List<CardAccountResponse> getCardList(Long memberId);
+}
