+++ src/main/java/com/team/independence/asset/service/LoanAccountService.java
@@ -1,0 +1,9 @@
+package com.team.independence.asset.service;
+
+import com.team.independence.asset.dto.account.LoanAccountDetailItem;
+
+import java.util.List;
+
+public interface LoanAccountService {
+    List<LoanAccountDetailItem> getLoanAccounts(Long memberId);
+}
